# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "nodejs-hw-mongodb",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "REST API for managing contacts and authentication.\n"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://nodejs-hw-mongodb-1-m7fq.onrender.com"
    }
  ],
...
  "paths": {
...
    "/contacts/{contactId}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get contact by ID",
        "operationId": "getContactById",
        "description": "Returns a contact by its unique ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the contact",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found contact",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contact with id 60d21b4667d0d8992e610c85!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                },
                "examples": {
                  "success": {
                    "value": {
                      "status": 200,
                      "message": "Successfully found contact with id 60d21b4667d0d8992e610c85!",
                      "data": {
                        "_id": "60d21b4667d0d8992e610c85",
                        "name": "John Doe",
                        "phoneNumber": "+1234567890",
                        "contactType": "work",
                        "photo": null,
                        "createdAt": "2025-09-11T15:24:13.523Z",
                        "updatedAt": "2025-09-11T15:24:13.523Z",
                        "__v": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Contacts"
        ],
        "summary": "Update contact by ID",
        "operationId": "updateContactById",
        "description": "Updates an existing contact by its unique ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": null,
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "+1234567890"
                  },
                  "contactType": {
                    "type": "string",
                    "example": "work"
                  },
                  "email": {
                    "type": "string",
                    "example": "emailexp@gmail.com"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "example": true
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary",
                    "description": "Contact photo file"
                  }
                }
              },
              "encoding": {
                "photo": {
                  "contentType": "image/png, image/jpeg"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully patched a contact!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                },
                "examples": {
                  "success": {
                    "value": {
                      "status": 200,
                      "message": "Successfully patched a contact!",
                      "data": {
                        "name": "John Doe",
                        "phoneNumber": "+1234567890",
                        "contactType": "work",
                        "userId": "68bc0c4a5113f30a9e70eb9f",
                        "photo": "https://res.cloudinary.com/dfhdxlg1y/image/upload/v1757605150/n672bhjkcm720amcxyu8.png",
                        "_id": "68c2ed1f979e3b1c8a78db94",
                        "createdAt": "2025-09-11T15:39:11.246Z",
                        "updatedAt": "2025-09-11T15:39:11.246Z",
                        "__v": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "summary": "Delete contact by ID",
        "operationId": "deleteContactById",
        "description": "Deletes a contact by its unique ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "description": "Unique identifier of the contact",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Contact deleted successfully. No content returned."
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "contact": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60d21b4667d0d8992e610c85"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+1234567890"
          },
          "contactType": {
            "type": "string",
            "example": "work"
          },
          "photo": {
            "type": "string",
            "format": "binary",
            "description": "Contact photo file"
          }
        },
        "required": [
          "name",
          "phoneNumber",
          "contactType"
        ]
      },
      "error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "Contact not found"
          }
        },
        "required": [
          "status",
          "message"
        ]
      }
    }
  }
}